feat(InteractionInvalidation): add options for partial invalidation (tabIndex and ariaHidden)
