+++ packages/dnb-eufemia/package.json
@@ -18,7 +18,7 @@
     "audit:ci": "yarn audit:ci:yarn",
     "audit:ci:npm": "audit-ci --config ./audit-ci.json --report-type summary",
     "audit:ci:yarn": "audit-ci --config ./audit-ci.json --package-manager=yarn --report-type summary",
-    "audit:ci:yarn-outdated": "babel-node ./scripts/prebuild/audit/toOpt && audit-ci --config ./audit-ci.json --package-manager=yarn --report-type full && babel-node ./scripts/prebuild/audit/toDev",
+    "audit:ci:yarn-outdated": "babel-node --extensions .js,.ts,.tsx ./scripts/prebuild/audit/toOpt && audit-ci --config ./audit-ci.json --package-manager=yarn --report-type full && babel-node --extensions .js,.ts,.tsx ./scripts/prebuild/audit/toDev",
     "build": "./scripts/prebuild/local-prebuild.sh",
     "prebuild:ci": "yarn build:prebuild",
     "build:ci": "yarn prebuild:ci && yarn postbuild:ci && yarn publish:prepare",
@@ -29,25 +29,25 @@
     "build:es": "./scripts/postbuild/babel-es.sh",
     "build:esm": "./scripts/postbuild/babel-esm.sh",
     "prebuild:figma:ci": "yarn figma",
-    "postbuild:commit": "babel-node ./scripts/figma/tasks/commitChanges",
+    "postbuild:commit": "babel-node --extensions .js,.ts,.tsx ./scripts/figma/tasks/commitChanges",
     "build:pack": "yarn build && yarn publish:prepare && cd ./build && yarn pack",
     "build:packages": "./scripts/postbuild/build-packages.sh",
-    "build:prebuild": "babel-node ./scripts/prebuild/runPrepub.js",
-    "build:resources": "babel-node ./scripts/prebuild/resources/makeResourcesPackage.js",
+    "build:prebuild": "babel-node --extensions .js,.ts,.tsx ./scripts/prebuild/runPrepub.js",
+    "build:resources": "babel-node --extensions .js,.ts,.tsx ./scripts/prebuild/resources/makeResourcesPackage.js",
     "build:types": "./scripts/prebuild/generate-types.sh",
     "build:types:definitions": "yarn tsc --project tsconfig.definitions.json",
-    "build:types:dev": "nodemon --exec 'babel-node ./scripts/prebuild/generateTypes.js' --ext js --watch './src/**/*' --watch './scripts/**/*'",
+    "build:types:dev": "nodemon --exec 'babel-node --extensions .js,.ts,.tsx ./scripts/prebuild/generateTypes.js' --ext js --watch './src/**/*' --watch './scripts/**/*'",
     "build:types:esm": "yarn tsc --project tsconfig.definitions.json --declarationDir ./build/esm --rootDir src",
     "precommit": "yarn lint-staged",
-    "dev:icons": "nodemon --exec 'babel-node ./scripts/tools/convertIcons' --ignore '/icons/**' --ignore '*.json'",
+    "dev:icons": "nodemon --exec 'babel-node --extensions .js,.ts,.tsx ./scripts/tools/convertIcons' --ignore '/icons/**' --ignore '*.json'",
     "dev:packages": "nodemon --exec 'yarn build:packages && yarn build:copy' --ext js,html,json,css,scss --watch './src/**/*' --ignore './umd/*'",
-    "dev:resources": "nodemon --exec 'babel-node ./scripts/prebuild/resources/makeResourcesPackage.js' --ext js,html,json,css,scss --watch './build/style/**/*' --watch './scripts/**/*' --ignore '*.json'",
-    "dev:tasks": "nodemon --exec 'babel-node ./scripts/prebuild/dev.js' --watch 'rollup.config.js' --ext js,html,json,css,scss --watch './src/components/**/*' --watch './src/style/**/*' --watch './scripts/**/*' --ignore '*.json'",
-    "figma": "babel-node ./scripts/figma/updateAll",
-    "figma:dev:commit": "nodemon --exec 'babel-node ./scripts/figma/tasks/commitChanges'",
-    "figma:dev:icons": "nodemon --exec 'babel-node ./scripts/figma/updateIcons' --ignore '*.svg' --ignore '*.json'",
+    "dev:resources": "nodemon --exec 'babel-node --extensions .js,.ts,.tsx ./scripts/prebuild/resources/makeResourcesPackage.js' --ext js,html,json,css,scss --watch './build/style/**/*' --watch './scripts/**/*' --ignore '*.json'",
+    "dev:tasks": "nodemon --exec 'babel-node --extensions .js,.ts,.tsx ./scripts/prebuild/dev.js' --watch 'rollup.config.js' --ext js,html,json,css,scss --watch './src/components/**/*' --watch './src/style/**/*' --watch './scripts/**/*' --ignore '*.json'",
+    "figma": "babel-node --extensions .js,.ts,.tsx ./scripts/figma/updateAll",
+    "figma:dev:commit": "nodemon --exec 'babel-node --extensions .js,.ts,.tsx ./scripts/figma/tasks/commitChanges'",
+    "figma:dev:icons": "nodemon --exec 'babel-node --extensions .js,.ts,.tsx ./scripts/figma/updateIcons' --ignore '*.svg' --ignore '*.json'",
     "figma:refetch": "yarn figma --force-refetch",
-    "figma:reset": "babel-node ./scripts/figma/resetIcons && babel-node ./scripts/tools/convertIcons",
+    "figma:reset": "babel-node --extensions .js,.ts,.tsx ./scripts/figma/resetIcons && babel-node --extensions .js,.ts,.tsx ./scripts/tools/convertIcons",
     "jest": "cross-env NODE_ICU_DATA=../../node_modules/full-icu jest",
     "lint": "eslint --quiet ./src ./scripts",
     "lint-staged": "lint-staged -d",
@@ -58,12 +58,12 @@
     "lint:lockfile": "yarn lockfile-lint --path ../../yarn.lock --type yarn --validate-https --allowed-hosts yarn github.com codeload.github.com",
     "lint:styles": "stylelint './src/**/*.{js,scss}'",
     "lint:styles:staged": "stylelint './{src,scripts}/**/*.{js,scss}'",
-    "make-properties": "babel-node ./scripts/prebuild/makeProperties",
+    "make-properties": "babel-node --extensions .js,.ts,.tsx ./scripts/prebuild/makeProperties",
     "publish:ci": "yarn publish:prepare && ./scripts/postbuild/publish-release.sh",
     "publish:dry": "dotenv semantic-release --no-ci --dry-run",
     "publish:prepare": "./scripts/postbuild/publish-prepare.sh",
     "reset": "rm -rf ./node_modules ./build",
-    "skeleton:font": "nodemon --exec 'babel-node ./scripts/tools/createSkeletonFont.js'",
+    "skeleton:font": "nodemon --exec 'babel-node --extensions .js,.ts,.tsx ./scripts/tools/createSkeletonFont.js'",
     "start": "start-storybook -s / -p 8002 --ci --no-release-notes --no-version-updates",
     "test": "jest",
     "test:auto-generated-types": "jest ./postTypeGeneration.test.js --ci --testPathIgnorePatterns=[]",
