+++ packages/dnb-eufemia/scripts/postbuild/babel-es.sh
@@ -19,5 +19,5 @@ echo 'Building es bundle done!'
 
 echo 'Copy additional files to es ...'
 
-OUT_DIR=./build/es babel-node ./scripts/postbuild/copyTypeScriptFiles.js
-OUT_DIR=./build/es babel-node ./scripts/postbuild/copyStyles.js
+OUT_DIR=./build/es babel-node --extensions .js,.ts,.tsx ./scripts/postbuild/copyTypeScriptFiles.js
+OUT_DIR=./build/es babel-node --extensions .js,.ts,.tsx ./scripts/postbuild/copyStyles.js
