+++ packages/dnb-eufemia/scripts/postbuild/babel-esm.sh
@@ -19,5 +19,5 @@ echo 'Building esm bundle done!'
 
 echo 'Copy additional files to esm ...'
 
-OUT_DIR=./build/esm babel-node ./scripts/postbuild/copyTypeScriptFiles.js
-OUT_DIR=./build/esm babel-node ./scripts/postbuild/copyStyles.js
+OUT_DIR=./build/esm babel-node --extensions .js,.ts,.tsx ./scripts/postbuild/copyTypeScriptFiles.js
+OUT_DIR=./build/esm babel-node --extensions .js,.ts,.tsx ./scripts/postbuild/copyStyles.js
