+++ packages/dnb-eufemia/scripts/prebuild/generate-types.sh
@@ -4,7 +4,7 @@ set -e # Exit immediately if a command exits with a non-zero status.
 
 echo 'Generating type definitions ...'
 
-babel-node ./scripts/prebuild/generateTypes.js
+babel-node --extensions .js,.ts,.tsx ./scripts/prebuild/generateTypes.js
 yarn build:types:definitions
 yarn test:auto-generated-types
